Optimize NumberRangePref call change
Reduce the cost of converting its value from a string to Int and vice-versa

This is done by calling `callChangeListener` manually. Previously, it was called on `super.OnDialogClosed()`, which passed the value as a string

+++ AnkiDroid/src/main/java/com/ichi2/preferences/NumberRangePreferenceCompat.kt
@@ -181,7 +181,7 @@ open class NumberRangePreferenceCompat @JvmOverloads constructor(
             if (!positiveResult || editText.text.isEmpty()) {
                 return
             }
-            val newValue = editText.text.toString()
+            val newValue = editText.text.toString().toInt()
             if (numberRangePreference.callChangeListener(newValue)) {
                 numberRangePreference.setValue(newValue)
             }
